Bench 6332093

+++ src/types.h
@@ -10,7 +10,7 @@
 // define bitboard data type
 using Bitboard = uint64_t;
 // define poskey data type
-using TTKey = uint16_t;
+using TTKey = uint64_t;
 // define poskey data type
 using ZobristKey = uint64_t;
 using Move = uint32_t;
